backend/requirements.txt
Django>=5.0,<6.0
Faker>=20.0.0,<30.0
djangorestframework>=3.15.0,<4.0
django-cors-headers>=4.4.0
psycopg[binary]>=3.2
python-dotenv>=1.0.1
drf-spectacular>=0.27.2
django-storages[s3]>=1.14.4
Pillow>=10.3.0
djangorestframework-simplejwt[crypto]>=5.3.1
boto3>=1.34.0
django-filter>=24.2
requests>=2.31.0
africastalking>=1.2.5
reportlab>=4.0.0
phonenumbers>=8.13.0
